move main view into view folder
