+++ App.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import MainView from "./components/MainView";
+import MainView from "./views/MainView";
 
 export default function App(): JSX.Element {
 	return <MainView />;
